backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
PyJWT==2.9.0
python-multipart==0.0.9
httpx==0.27.2
boto3==1.35.24
passlib[bcrypt]==1.7.4
python-telegram-bot==21.6
